Separated out the main Tkinter object, previously _App and now AdministratorWindow, into it's own file.
